Integrated cosmos playground.
