+++ fixtures/CheckboxField/default.js
@@ -1,0 +1,4 @@
+module.exports = {
+  schema: {},
+  label: "foo",
+};
