+++ webpack.config.dist.js
